Update broken sync tests
Fixes sync tests broken by sync error refactors and migrates to
`sync/connection.rs`.
